examples/requirements.txt
streamlit!=1.24.0
streamlit-folium>=0.11.1
requests
geopandas